git-svn-id: svn://localhost/ardour2/branches/3.0@3711 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/automation_line.h
@@ -94,7 +94,7 @@ class AutomationLine : public sigc::trackable, public PBD::StatefulThingWithGoin
 
 	void    show ();
 	void    hide ();
-	void    set_y_position_and_height (uint32_t, uint32_t);
+	void    set_y_position_and_height (double, double);
 	void    set_verbose_cursor_uses_gain_mapping (bool yn);
 
 	TimeAxisView& trackview;
